Create week-1&2-fibname.py
Comments from the forum for week 1:
My name is Nuala, so the first and last letter of my name (N + A = 14+1) give the number 15. The 15th Fibonacci number is 610.

For Fionnuala, the first and last letter of my name is F+A = 6+1, which give the number 7. The 7th Fibonacci number is 13.


Comments from the forum for week 2:
My surname is Gallagher

The first letter G is number 71

The last letter r is number 114

Fibonacci number 185 is 205697230343233228174223751303346572685

The ord() method returns an integer representing Unicode code point for the given Unicode character.